+++ src/iban/iban.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'bun:test';
-import { calculateCheckDigits, getIban } from './iban';
+import { calculateCheckDigits, normalizedAccountToIban } from './iban';
 
 describe('iban', () => {
   it('calculateCheckDigits', () => {
@@ -12,9 +12,9 @@ describe('iban', () => {
     ).toEqual('43');
   });
 
-  it('getIban', () => {
+  it('normalizedAccountToIban', () => {
     expect(
-      getIban({
+      normalizedAccountToIban({
         bankCode: '3030',
         prefix: '000000',
         number: '1263035066',
@@ -27,7 +27,7 @@ describe('iban', () => {
     ['2010', '000000', '2901360000', 'CZ2220100000002901360000'],
     ['0100', '000123', '1234567891', 'CZ5701000001231234567891'],
     ['0800', '000019', '2000145399', 'CZ6508000000192000145399'],
-  ])('getIban for %s/%s/%s', (bankCode, prefix, number, expected) => {
-    expect(getIban({ bankCode, prefix, number })).toEqual(expected);
+  ])('normalizedAccountToIban for %s/%s/%s', (bankCode, prefix, number, expected) => {
+    expect(normalizedAccountToIban({ bankCode, prefix, number })).toEqual(expected);
   });
 });